Show warning before deleting features
git-svn-id: http://svn.osgeo.org/qgis/trunk/qgis@10165 c8812cc2-4d05-0410-92ff-de0c093fc19c

+++ src/app/qgisapp.cpp
@@ -3749,6 +3749,14 @@ void QgisApp::deleteSelected()
     return;
   }
 
+  //display a warning
+  int numberOfDeletedFeatures = vlayer->selectedFeaturesIds().size();
+  if(QMessageBox::warning(this, tr("Delete features"), tr("Delete %1 feature(s)?").arg(numberOfDeletedFeatures), QMessageBox::Ok, QMessageBox::Cancel) == QMessageBox::Cancel)
+  {
+    return;
+  }
+
+
   if ( !vlayer->deleteSelectedFeatures() )
   {
     QMessageBox::information( this, tr( "Problem deleting features" ),
